+++ lib/td_dd_web/controllers/data_structure_controller.ex
@@ -397,7 +397,7 @@ defmodule TdDdWeb.DataStructureController do
     params
     |> Map.put("without", "deleted_at")
     |> Map.drop(["page", "size"])
-    |> Search.search_data_structures(claims, permission, 0, :infinity)
+    |> Search.scroll_data_structures(claims, permission)
   end
 
   swagger_path :csv do