+++ lib/td_dd_web/schema.ex
@@ -9,9 +9,11 @@ defmodule TdDdWeb.Schema do
 
   import_types(Absinthe.Plug.Types)
   import_types(Absinthe.Type.Custom)
+  import_types(TdDdWeb.Schema.DataStructureLinks)
   import_types(TdDdWeb.Schema.Domains)
   import_types(TdDdWeb.Schema.Executions)
   import_types(TdDdWeb.Schema.Functions)
+  import_types(TdDdWeb.Schema.GrantApprovalRules)
   import_types(TdDdWeb.Schema.GrantRequests)
   import_types(TdDdWeb.Schema.ImplementationResults)
   import_types(TdDdWeb.Schema.Implementations)
@@ -32,6 +34,7 @@ defmodule TdDdWeb.Schema do
     import_fields(:domain_queries)
     import_fields(:function_queries)
     import_fields(:grant_request_queries)
+    import_fields(:grant_approval_rules_queries)
     import_fields(:implementation_queries)
     import_fields(:implementation_results_queries)
     import_fields(:me_queries)
@@ -45,6 +48,7 @@ defmodule TdDdWeb.Schema do
   end
 
   mutation do
+    import_fields(:grant_approval_rules_mutations)
     import_fields(:implementation_mutations)
     import_fields(:reference_data_mutations)
     import_fields(:source_mutations)