+++ lib/td_dd_web/schema/structures.ex
@@ -53,7 +53,7 @@ defmodule TdDdWeb.Schema.Structures do
     field :confidential, non_null(:boolean)
     field :domain_id, :integer, resolve: &Resolvers.Structures.domain_id/3
     field :domain_ids, list_of(:integer)
-    field :domains, list_of(:domain), resolve: &Resolvers.Structures.domains/3
+    field :domains, list_of(:domain), resolve: &Resolvers.Domains.domains/3
     field :external_id, non_null(:string)
     field :inserted_at, :datetime
     field :updated_at, :datetime
@@ -63,6 +63,9 @@ defmodule TdDdWeb.Schema.Structures do
 
     field :structure_tags, list_of(:structure_tag),
       resolve: &Resolvers.Structures.structure_tags/3
+
+    field :data_structure_links, list_of(:data_structure_link),
+      resolve: &Resolvers.Structures.data_structure_links/3
 
     field :available_tags, list_of(:tag), resolve: &Resolvers.Structures.available_tags/3
   end
@@ -82,10 +85,18 @@ defmodule TdDdWeb.Schema.Structures do
     field :data_structure, :data_structure, resolve: dataloader(TdDd.DataStructures)
     field :path, list_of(:string), resolve: &Resolvers.Structures.data_structure_version_path/3
 
+    field :path_with_ids, list_of(:path_with_id),
+      resolve: &Resolvers.Structures.data_structure_version_path_with_ids/3
+
     field :parents, list_of(:data_structure_version) do
       arg(:deleted, :boolean, default_value: false)
       resolve(dataloader(TdDd.DataStructures))
     end
+  end
+
+  object :path_with_id do
+    field :data_structure_id, non_null(:id)
+    field :name, :string
   end
 
   object :data_structure_relation do