+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [TD-5289] Elasticsearch 7 compatibility
+- [TD-5342] Refactor bulk operations to use Elasticsearch scroll API
+- [TD-5136] Download of structures include all structure domains with complete
+  path
+
+### Added
+
+- [TD-5258] Approve grants request rules
+
+## [4.55.0] 2022-11-14
+
+### Added
+
+- [TD-5038] Links between structures
+- [TD-5037] Added diff information for draft and pending_aproval structure notes
+  in API response
+
 ### Fixed
 
 - [TD-5279] Fixed some inconsistencies in implementation bulk upload